rename jwt.py to jwt_helper.py to fix a name conflict with the jwt library
