+++ qhana_plugin_runner/api/__init__.py
@@ -28,7 +28,7 @@ from flask_smorest import Blueprint as SmorestBlueprint
 
 from .extra_fields import EnumField
 from .files_api import FILES_API
-from .jwt import SECURITY_SCHEMES
+from .jwt_helper import SECURITY_SCHEMES
 from .plugins_api import PLUGINS_API
 from .tasks_api import TASKS_API
 from .util import MaBaseSchema
